backend/requirements.txt
# Python 3.12.4 Compatible Requirements
# Web framework
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.3
pydantic-settings==2.2.1

# ML packages - Python 3.12 compatible versions with tokenizer fix
torch==2.2.1
transformers==4.38.2
sentence-transformers==2.6.1
scikit-learn==1.4.1.post1
numpy==1.26.4
huggingface-hub==0.21.4
accelerate==0.27.2
tokenizers>=0.15.2

# HTTP and utilities
httpx==0.27.0
requests==2.31.0
python-multipart==0.0.9

# Document processing
PyPDF2==3.0.1
pdfplumber==0.10.4
python-docx==1.1.0
docx2txt==0.8
python-magic-bin==0.4.14

# Database
beanie==1.26.0
motor==3.4.0
pymongo==4.6.2

# Background tasks
celery==5.3.6
redis==5.0.3

# Auth
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
cryptography==42.0.5
passlib[bcrypt]==1.7.4

# AI/LLM - Updated versions
langchain==0.1.13
openai==1.14.3

# Rate limiting
slowapi==0.1.9

# Pydantic extras
pydantic[email]==2.6.3